Bassim/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
openai==1.14.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
